photoevent/services/ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch==2.1.2
torchvision==0.16.2
facenet-pytorch==2.5.3
Pillow==10.2.0
numpy==1.26.3
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.6
